Esta terminada la logica del form de servir cerveza. Falta hacer el resumen del negocio y corregir algunos errores en la actualizacion de datos en el formulario: frm_servirCerveza

+++ guia_ejercicios/ejercicio04/frm_agregarBarril.cs
@@ -51,7 +51,7 @@ namespace ejercicio04
                 {
                     Cerveza cervezaElegida = cervezas_comboBox.SelectedItem as Cerveza;
 
-                    Barril nuevoBarril = new Barril((int)id_numericUpDown.Value, cervezaElegida, (int)capacidad_numericUpDown.Value);
+                    Barril nuevoBarril = new Barril((int)id_numericUpDown.Value, cervezaElegida, (float)capacidad_numericUpDown.Value);
                     afterOffice.AgregarBarril(nuevoBarril);
                     this.Close();
                 } else
